backend/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.8
asyncpg==0.29.0
aiosqlite==0.20.0
beautifulsoup4==4.10.0
black==24.4.0
cachetools==5.3.2
click==8.1.7
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.110.1
fastkml==0.12
greenlet==3.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.1
pydantic_core==2.10.1
pygeoif==0.7
pylint==2.17.6
pymemcache==4.0.0
pytest==7.4.3
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-memcached==1.59
python-multipart==0.0.7
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.1
types-cachetools==5.3.0.7
types-psycopg2==2.9.21.14
typing_extensions==4.8.0
uvicorn==0.23.2
websockets==12.0
wrapt==1.15.0
